testsprite_tests/requirements.txt
requests>=2.31.0
PyJWT>=2.8.0
pytest>=7.4.0

